requirements.txt
discord.py
pyarango